test: consolidate TestConsistency
