+++ internal/services/integrationtesting/consistencytestutil/clusteranddata.go
@@ -3,21 +3,14 @@ package consistencytestutil
 import (
 	"context"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
 
-	"github.com/authzed/spicedb/internal/datastore/dsfortesting"
-	"github.com/authzed/spicedb/internal/datastore/memdb"
 	"github.com/authzed/spicedb/internal/dispatch"
 	"github.com/authzed/spicedb/internal/dispatch/caching"
 	"github.com/authzed/spicedb/internal/dispatch/graph"
 	"github.com/authzed/spicedb/internal/dispatch/keys"
-	"github.com/authzed/spicedb/internal/testserver"
-	caveattypes "github.com/authzed/spicedb/pkg/caveats/types"
-	"github.com/authzed/spicedb/pkg/cmd/server"
-	"github.com/authzed/spicedb/pkg/datalayer"
 	"github.com/authzed/spicedb/pkg/datastore"
 	"github.com/authzed/spicedb/pkg/validationfile"
 )
@@ -29,40 +22,6 @@ type ConsistencyClusterAndData struct {
 	DataStore datastore.Datastore
 	Ctx       context.Context
 	Populated *validationfile.PopulatedValidationFile
-}
-
-// LoadDataAndCreateClusterForTesting loads the data found in a consistency test file,
-// builds a cluster for it, and returns both the data and cluster.
-func LoadDataAndCreateClusterForTesting(t *testing.T, consistencyTestFilePath string, revisionDelta time.Duration, additionalServerOptions ...server.ConfigOption) ConsistencyClusterAndData {
-	require := require.New(t)
-
-	ds, err := dsfortesting.NewMemDBDatastoreForTesting(t, 0, revisionDelta, memdb.DisableGC)
-	require.NoError(err)
-
-	return BuildDataAndCreateClusterForTesting(t, consistencyTestFilePath, ds, additionalServerOptions...)
-}
-
-// BuildDataAndCreateClusterForTesting loads the data found in a consistency test file,
-// builds a cluster for it, and returns both the data and cluster.
-func BuildDataAndCreateClusterForTesting(t *testing.T, consistencyTestFilePath string, ds datastore.Datastore, additionalServerOptions ...server.ConfigOption) ConsistencyClusterAndData {
-	require := require.New(t)
-
-	populated, _, err := validationfile.PopulateFromFiles(t.Context(), datalayer.NewDataLayer(ds), caveattypes.Default.TypeSet, []string{consistencyTestFilePath})
-	require.NoError(err)
-
-	connections := testserver.TestClusterWithDispatch(t, 1, ds, additionalServerOptions...)
-
-	dl := datalayer.NewDataLayer(ds)
-
-	dsCtx := datalayer.ContextWithHandle(t.Context())
-	require.NoError(datalayer.SetInContext(dsCtx, dl))
-
-	return ConsistencyClusterAndData{
-		Conn:      connections[0],
-		DataStore: ds,
-		Ctx:       dsCtx,
-		Populated: populated,
-	}
 }
 
 // CreateDispatcherForTesting creates a dispatcher for consistency testing, with or without
